springapi/app: Add prod/dev/test environments to config
Moved config.py from top level to springapi/ because it only affects this
part of the app.

During tests, the app has the test configuration applied (because it is
testing something). This happens in the test itself, but other configs
could be applied by passing command line args to pytest.

resolving #5