models.Firebase: Successful PUT/POST return 200, exclude ID field
PUT/POST requests remove the ID field before request touches Firebase API
so to avoid creating an unwanted field. POST requests now return 200 and
a JSON object of the created entry when successful (had previously been
returning 500 on success).

A big part of the problem was handling the ID field as it was passed to
Firebase model. Solved this by passing entry object using copy(), which
solves the issue but is difficult ot test. Tests do not check for use of
copy as of this commit (but really should).

resolves #14