models.submission: Add _create_uid()
Creates a UID for new submissions. Called by create_submission(). ID no
longer belongs in body of request and will be overwritten (POST and PUT).

It was very difficult to write a test which confirms that this method
creates an ID because the ID needs to be known by the test before one is
generated. This is a gap in test coverage.

resolves #15