WIP CI: Add build stage, registers Docker image to AWS ECR
Everything needed is here, except Docker is not installed in runner so it
is not possible for the pipeline to complete during build. Official
documentation states that docker image should be used, but then test
stage won't work because Python is missing. Both Python and Docker are
heavy installs, don't want to guess-and-test this using live resources.