users: Get list of users who can have access to API
Gets a list of users from some database, currently Firebase for
simplicity's sake. Requires the USERS environment variable to be set with
a protocol indicating the database and credentials (in JSON) to access
that database.

The check for the presence of a firebase_admin app is to avoid double
authentication, as Firebase is also used for Submission model.

models.firebase.app: Rename to models.firebase.authenticate to avoid
confusion with newly added firebase modules.

helpers: VALID_USERS -> TOKENS to avoid namespace conflict with USERS