auth: Remove TOKEN environment variable, replace with User
User contains the token of a user, if set. As of this commit, there is no
way to set the token using this app (that is part of next step). But the
mechanic now exists to check the bearer token by a list of tokens which
are associated with valid, authorised users.