routes.auth: Fix token exchange route
Had previously contained incorrect body and skipped credentials (among
other things); also used GET instead of POST. Corrected this, but have
not yet implemented anything which would handle token storage. The
response to client is simply the token, which is not desireable. These
are to be address in next commit.