auth: Check authenticated user email address against valid email list
After the user has passed authorisation from an OAuth provider, his
associated email address is checked by a list of valid emails to confirm
identity.

As of this commit, the token from the OAuth provider is returned to the
requesting application. This will shortly be replaced with a token
generated by the API.