auth: Move to utils/, refactor app, move global variables
Auth never belonged in models, so created utils/ and moved it there.
Keeping utils/authorization agnostic to OAuth provider.

The way that configuration was being passed to auth (to generate OAuth
request URI) was convoluted. Simplified this, and in doing so refactored
much of springapi.app.

Moved anything not involved in database or app
creation to config helpers. Also moved global variables here.