auth: Create, store, and return to requester API token; key pending
API token is generated by PyJWT and contains only iss and email. There is
no expiry date as of yet but one could be added later. The key is just a
string for now, will move to env variable in next commit.

This commit effectively makes User model obsolete. It should be removed
from the app shortly.